+++ Source/Objects/Hardware/Pci/A3818/ORA3818Model.m
@@ -824,7 +824,7 @@ NSString* ORA3818Lock										= @"ORA3818Lock";
 			//if(*fVStatusReg & STATUS_PROBLEM)[self checkStatusWord:status];
 			// transfer data
 			uint32_t *pulr = (uint32_t *)( ( vmeAddress & 0x00000fff ) +
-													(uint32_t)remMemRegisterAddress );
+													remMemRegisterAddress );
 			uint32_t *pulb = (uint32_t *)readAddress;
 			unsigned int n = numberLongs;
 			for(;n--;)*pulb++ = *pulr++;
@@ -860,7 +860,7 @@ NSString* ORA3818Lock										= @"ORA3818Lock";
             
             
 			// transfer data
-			uint32_t *pulr = (uint32_t *)( ( vmeAddress & 0x00000fff ) + (uint32_t)remMemRegisterAddress );
+			uint32_t *pulr = (uint32_t *)( ( vmeAddress & 0x00000fff ) + remMemRegisterAddress );
 			uint32_t *pulb = (uint32_t *)readAddress;
 			unsigned int n = numberLongs;
 			for(;n--;)*pulb++ = *pulr++;
@@ -903,7 +903,7 @@ NSString* ORA3818Lock										= @"ORA3818Lock";
 			//if(*fVStatusReg & STATUS_PROBLEM)[self checkStatusWord:status];
 			// transfer data
 			uint32_t *pulr = (uint32_t *)( ( vmeAddress & 0x00000fff ) +
-													(uint32_t)remMemRegisterAddress );
+													remMemRegisterAddress );
 			uint32_t *pulb = (uint32_t *)writeAddress;
 			unsigned int n = numberLongs;
 			for(;n--;)*pulr++ = *pulb++;
@@ -1010,7 +1010,7 @@ typedef struct a3818_comm {
 			//if(*fVStatusReg & STATUS_PROBLEM)[self checkStatusWord:*fVStatusReg];
 			// transfer data
 			unsigned char *pulr = (unsigned char *)( ( vmeAddress & 0x00000fff ) +
-													(uint32_t)remMemRegisterAddress );
+													remMemRegisterAddress );
 			unsigned char *pulb = (unsigned char *)writeAddress;
 			unsigned int n = numberBytes;
 			for(;n--;)*pulr++ = *pulb++;
@@ -1047,7 +1047,7 @@ typedef struct a3818_comm {
 			//if(*fVStatusReg & STATUS_PROBLEM)[self checkStatusWord:*fVStatusReg];
             // transfer data
 			uint16_t *pulr = (uint16_t *)( ( vmeAddress & 0x00000fff ) +
-									  (uint32_t)remMemRegisterAddress );
+									  remMemRegisterAddress );
 			uint16_t *pulb = (uint16_t *)readAddress;
 			unsigned short n = numberWords;
 			for(;n--;)*pulb++ = *pulr++;
@@ -1087,7 +1087,7 @@ typedef struct a3818_comm {
 			//if(*fVStatusReg & STATUS_PROBLEM)[self checkStatusWord:*fVStatusReg];
 			// transfer data
 			uint16_t *pulr = (uint16_t *)( ( vmeAddress & 0x00000fff ) +
-									  (uint32_t)remMemRegisterAddress );
+									  remMemRegisterAddress );
 			uint16_t *pulb = (uint16_t *)writeAddress;
 			unsigned int n = numberWords;
 			for(;n--;)*pulr++ = *pulb++;
@@ -1765,8 +1765,8 @@ physicalBufferAddress:(uint32_t) physicalBufferAddress
 	//	volatile uint32_t *mptr = (uint32_t *)( (uint32_t)GetMappingBaseAddress() +
 	//			 (uint32_t)DMA_MAPPING_REGISTER_OFFSET );
 	
-	volatile uint32_t *mptr = (uint32_t *)( (uint32_t)mapRegisterAddress +
-													 (uint32_t)DMA_MAPPING_REGISTER_OFFSET );
+	volatile uint32_t *mptr = (uint32_t *)(mapRegisterAddress +
+													 DMA_MAPPING_REGISTER_OFFSET );
 	
 	
 	uint32_t j;
